# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Group Broadcaster API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WhatsappStatusState =
  (typeof WhatsappStatusState)[keyof typeof WhatsappStatusState];

export const WhatsappStatusState = {
  initializing: "initializing",
  qr_ready: "qr_ready",
  pairing: "pairing",
  authenticated: "authenticated",
  ready: "ready",
  disconnected: "disconnected",
} as const;

export interface WhatsappStatus {
  state: WhatsappStatusState;
  /** Base64 data URL of the QR code image when state is qr_ready */
  qrCode?: string | null;
  /** 8-character pairing code when state is pairing */
  pairingCode?: string | null;
  phoneNumber?: string | null;
  name?: string | null;
  groupCount?: number | null;
}

export interface WhatsappGroup {
  id: string;
  name: string;
  participantCount: number;
}

export interface GroupsResponse {
  groups: WhatsappGroup[];
}

export interface BroadcastRequest {
  /** The message to broadcast to all groups */
  message: string;
  /** Optional list of specific group IDs to send to. If null, sends to all groups. */
  groupIds?: string[] | null;
}

export interface GroupSendResult {
  groupId: string;
  groupName: string;
  success: boolean;
  messageId?: string | null;
  error?: string | null;
  /** True if message received double ticks (delivered to server) */
  acked?: boolean | null;
}

export interface BroadcastResult {
  total: number;
  sent: number;
  failed: number;
  results: GroupSendResult[];
}

export interface DisconnectResult {
  success: boolean;
  message: string;
}