# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WhatsApp Group Broadcaster API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/broadcast:
    post:
      operationId: broadcastMessage
      tags: [whatsapp]
      summary: Broadcast message to all groups
      description: Sends a message to all WhatsApp groups and tracks delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastRequest"
      responses:
        "200":
          description: Broadcast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
...
components:
  schemas:
...
    BroadcastRequest:
      type: object
      properties:
        message:
          type: string
          description: The message to broadcast to all groups
        groupIds:
          type: array
          items:
            type: string
          nullable: true
          description: Optional list of specific group IDs to send to. If null, sends to all groups.
      required:
        - message
    BroadcastResult:
      type: object
      properties:
        total:
          type: integer
        sent:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/GroupSendResult"
      required:
        - total
        - sent
        - failed
        - results
    GroupSendResult:
      type: object
      properties:
        groupId:
          type: string
        groupName:
          type: string
        success:
          type: boolean
        messageId:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        acked:
          type: boolean
          nullable: true
          description: True if message received double ticks (delivered to server)
      required:
        - groupId
        - groupName
        - success
...