# lib/api-zod/src/generated/types/broadcastRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Group Broadcaster API
 * OpenAPI spec version: 0.1.0
 */

export interface BroadcastRequest {
  /** The message to broadcast to all groups */
  message: string;
  /** Optional list of specific group IDs to send to. If null, sends to all groups. */
  groupIds?: string[] | null;
}