# lib/api-zod/src/generated/types/broadcastResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Group Broadcaster API
 * OpenAPI spec version: 0.1.0
 */
import type { GroupSendResult } from "./groupSendResult";

export interface BroadcastResult {
  total: number;
  sent: number;
  failed: number;
  results: GroupSendResult[];
}