# lib/api-zod/src/generated/types/groupSendResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Group Broadcaster API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupSendResult {
  groupId: string;
  groupName: string;
  success: boolean;
  messageId?: string | null;
  error?: string | null;
  /** True if message received double ticks (delivered to server) */
  acked?: boolean | null;
}